clean code and file separation
